chore(api): update RQ-dashboard dependency and improve devcontainer setup
- Changed RQ-dashboard dependency from a Git URL to a PyPI source in multiple configuration files for consistency.
- Enhanced devcontainer configuration by adding host mapping and forwarding ports for better development experience.
- Updated Dockerfile to include docker-compose-plugin for improved container management.
- Adjusted task scripts to reference the correct path for Redis wait script and added a new script for cleaning stale RQ workers.
- Made database migration scripts more robust by checking for existing columns before adding or dropping them.

+++ projects/ecoindex_api/alembic/versions/5afa2faea43f_.py
@@ -8,6 +8,7 @@ Create Date: 2025-01-14 14:12:47.013413
 import sqlalchemy as sa
 import sqlmodel
 from alembic import op
+from ecoindex.database.helper import column_exists
 
 revision = "5afa2faea43f"
 down_revision = "7eaafaa65b32"
@@ -16,11 +17,13 @@ depends_on = None
 
 
 def upgrade() -> None:
-    op.add_column(
-        "apiecoindex",
-        sa.Column("source", sqlmodel.sql.sqltypes.AutoString(), nullable=True),  # type: ignore
-    )
+    if not column_exists(op.get_bind(), "apiecoindex", "source"):
+        op.add_column(
+            "apiecoindex",
+            sa.Column("source", sqlmodel.sql.sqltypes.AutoString(), nullable=True),  # type: ignore
+        )
 
 
 def downgrade() -> None:
-    op.drop_column("apiecoindex", "source")
+    if column_exists(op.get_bind(), "apiecoindex", "source"):
+        op.drop_column("apiecoindex", "source")
